+++ mesh/test/TestMemfemMeshReader.hpp
@@ -28,9 +28,9 @@ public:
             pMeshReader = new READER_3D(
                               "mesh/test/data/Memfem_slab"));
                               
-        TS_ASSERT(pMeshReader->GetNumNodes() == 381);
-        TS_ASSERT(pMeshReader->GetNumElements() == 1030);
-        TS_ASSERT(pMeshReader->GetNumFaces() == 758);
+        TS_ASSERT(pMeshReader->GetNumNodes() == 381U);
+        TS_ASSERT(pMeshReader->GetNumElements() == 1030U);
+        TS_ASSERT(pMeshReader->GetNumFaces() == 758U);
         
         std::vector<unsigned> NextFace;
         
@@ -40,7 +40,7 @@ public:
         TS_ASSERT( NextFace[1] == 23 );
         TS_ASSERT( NextFace[2] == 374 );
         
-        TS_ASSERT(pMeshReader->GetMaxNodeIndex() == (int) pMeshReader->GetNumNodes() - 1);
+        TS_ASSERT(pMeshReader->GetMaxNodeIndex() == pMeshReader->GetNumNodes() - 1);
         
         TS_ASSERT(pMeshReader->GetMinNodeIndex() == 0);
